feat: make each table its own page
